move canvasUtils into columnUtils
