--- packages/react-data-grid/src/utils/canvasUtils.ts
@@ -1,34 +1,0 @@
-import { isFrozen } from './columnUtils';
-import { CalculatedColumn } from '../common/types';
-
-export function getColumnScrollPosition<R>(columns: CalculatedColumn<R>[], idx: number, currentScrollLeft: number, currentClientWidth: number): number {
-  let left = 0;
-  let frozen = 0;
-
-  for (let i = 0; i < idx; i++) {
-    const column = columns[i];
-    if (column) {
-      if (column.width) {
-        left += column.width;
-      }
-      if (isFrozen(column)) {
-        frozen += column.width;
-      }
-    }
-  }
-
-  const selectedColumn = columns[idx];
-  if (selectedColumn) {
-    const scrollLeft = left - frozen - currentScrollLeft;
-    const scrollRight = left + selectedColumn.width - currentScrollLeft;
-
-    if (scrollLeft < 0) {
-      return scrollLeft;
-    }
-    if (scrollRight > currentClientWidth) {
-      return scrollRight - currentClientWidth;
-    }
-  }
-
-  return 0;
-}
